requirements.txt
# Core dependencies for Vercel Functions
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4

# Google APIs
pytrends==4.9.2
google-api-python-client==2.140.0
google-auth==2.32.0
google-auth-httplib2==0.2.0

# Web scraping
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.2.2
regex==2024.5.15

# HTTP & async
requests==2.32.3
httpx==0.27.0
aiohttp==3.9.5

# Data processing
pytz==2024.1
python-dateutil==2.9.0

# Monitoring & logging
structlog==24.2.0
colorama==0.4.6

# Analysis
matplotlib==3.8.4
seaborn==0.13.2